+++ example/lib/main.dart
@@ -14,13 +14,13 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
         visualDensity: VisualDensity.adaptivePlatformDensity,
       ),
-      home: MyHomePage(title: 'Flutter Demo Home Page'),
+      home: const MyHomePage(title: 'Flutter Demo Home Page'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
+  const MyHomePage({Key key, this.title}) : super(key: key);
 
   final String title;
 
@@ -39,9 +39,14 @@ class _MyHomePageState extends State<MyHomePage> {
           child: TextChunkStyling(
             text:
                 'Lorem ipsum dolor sit amet, consectetur adipiscing elit. Vivamus erat purus, sollicitudin et velit sed, suscipit euismod ipsum. Etiam ultricies purus a nunc condimentum, sollicitudin mollis tortor maximus. Phasellus fringilla augue a leo molestie feugiat. Donec eget nisi vel metus rhoncus ultricies. Donec non semper mi. Suspendisse dictum orci molestie libero vehicula, ut faucibus massa luctus. Etiam sit amet urna tristique, ullamcorper ex at, feugiat ipsum. Mauris ut leo quis magna tempus euismod. Nam euismod mauris quam, quis iaculis ligula ornare quis. Nunc egestas urna ac mauris consequat, id tincidunt justo iaculis. Ut posuere risus elit, vel facilisis nulla lobortis non.',
-            highlightText: ['sum', 'ing', 'mod', 'ris', 'nam'],
-            highlightTextStyle:
-                TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
+            highlightText: const ['sum', 'ing', 'mod', 'ris', 'nam'],
+            multiTextStyles: const [
+              TextStyle(color: Colors.red, fontWeight: FontWeight.bold),
+              TextStyle(color: Colors.amber, fontWeight: FontWeight.bold),
+              TextStyle(color: Colors.blue, fontWeight: FontWeight.bold),
+              TextStyle(color: Colors.brown, fontWeight: FontWeight.bold),
+              TextStyle(color: Colors.cyan, fontWeight: FontWeight.bold),
+            ],
           ),
         ),
       ),