+++ lib/src/text_chunk_styling.dart
@@ -19,8 +19,15 @@ class TextChunkStyling extends StatelessWidget {
   /// replace the closest enclosing [DefaultTextStyle].
   final TextStyle? textStyle;
 
-  /// Style to apply to the highlighted text.
-  final TextStyle highlightTextStyle;
+  /// Style to apply to the all elements of [highlightText].
+  final TextStyle? highlightTextStyle;
+
+  /// Multiple styles to apply to the text.
+  /// Each element in this list correspond to the style to apply at the same
+  /// index of [highlightText].
+  ///
+  /// This list must be the same length as [highlightText].
+  final List<TextStyle> multiTextStyles;
 
   /// How the text should be aligned horizontally.
   final TextAlign textAlign;
@@ -30,7 +37,8 @@ class TextChunkStyling extends StatelessWidget {
 
   /// Whether the text should break at soft line breaks.
   ///
-  /// If false, the glyphs in the text will be positioned as if there was unlimited horizontal space.
+  /// If false, the glyphs in the text will be positioned as if there was
+  /// unlimited horizontal space.
   final bool softWrap;
 
   /// The directionality of the text.
@@ -83,7 +91,8 @@ class TextChunkStyling extends StatelessWidget {
     required List<String> highlightText,
     this.caseSensitive = true,
     this.textStyle,
-    required this.highlightTextStyle,
+    this.highlightTextStyle,
+    this.multiTextStyles = const [],
     this.textAlign = TextAlign.start,
     this.overflow = TextOverflow.clip,
     this.softWrap = true,
@@ -95,50 +104,60 @@ class TextChunkStyling extends StatelessWidget {
     this.strutStyle,
   })  : assert(highlightText.isNotEmpty),
         assert(maxLines == null || maxLines > 0),
-        this.highlightText = highlightText
+        assert(
+          highlightTextStyle == null || multiTextStyles.isEmpty,
+          'Cannot provide both highlightTextStyle and multiTextStyles',
+        ),
+        highlightText = highlightText
             .map((e) => caseSensitive ? e : e.toLowerCase())
             .toList(),
-        super(key: key);
+        super(key: key) {
+    if (multiTextStyles.isNotEmpty) {
+      assert(multiTextStyles.length == highlightText.length);
+    }
+  }
 
   @override
   Widget build(BuildContext context) {
     assert(textDirection != null || debugCheckHasDirectionality(context));
 
-    // Define used TextStyle for classicText.
+    // Define used TextStyle for non-highlighted text.
     final defaultTextStyle = DefaultTextStyle.of(context);
     TextStyle? effectiveTextStyle = textStyle;
-    if (textStyle == null || textStyle!.inherit)
+    if (textStyle == null || textStyle!.inherit) {
       effectiveTextStyle = defaultTextStyle.style.merge(textStyle);
+    }
+
+    // Define used TextStyle for highlighted text.
+    final _highlightTextStyles = highlightTextStyle != null
+        ? List<TextStyle>.generate(
+            highlightText.length, (_) => highlightTextStyle!)
+        : multiTextStyles;
 
     final pattern =
-        RegExp('${highlightText.join("|")}', caseSensitive: caseSensitive);
-    final startWithPattern = text.startsWith(pattern);
-    final classicText = text.split(pattern)..removeWhere((e) => e.isEmpty);
+        RegExp(highlightText.join("|"), caseSensitive: caseSensitive);
 
-    // Create a string separator to isolate the text to highlight.
-    String strSeparator = '*';
-    while (text.contains(strSeparator)) strSeparator += '*';
-    final textToHighlight = text
-        .splitMapJoin(pattern,
-            onMatch: (m) => '${m.group(0)}', onNonMatch: (_) => strSeparator)
-        .split(strSeparator)
-          ..removeWhere((e) => e.isEmpty);
-
-    // Combine the two list in correct order
-    final combined = <String>[];
-    if (startWithPattern) {
-      for (int i = 0; i < textToHighlight.length; i++) {
-        if (i < textToHighlight.length)
-          combined.add(textToHighlight.elementAt(i));
-        if (i < classicText.length) combined.add(classicText.elementAt(i));
+    final _texts = <TextSpan>[];
+    var i = 0, j = 0;
+    while ((j = text.indexOf(pattern, i)) != -1) {
+      if (j > i) _texts.add(TextSpan(text: text.substring(i, j)));
+      final currentText = text.substring(j);
+      late final String what;
+      for (final e in highlightText) {
+        if (currentText.startsWith(RegExp(e, caseSensitive: caseSensitive))) {
+          what = e;
+          break;
+        }
       }
-    } else {
-      for (int i = 0; i < classicText.length; i++) {
-        if (i < classicText.length) combined.add(classicText.elementAt(i));
-        if (i < textToHighlight.length)
-          combined.add(textToHighlight.elementAt(i));
-      }
+      _texts.add(
+        TextSpan(
+          text: text.substring(j, j + what.length),
+          style: _highlightTextStyles[highlightText.indexOf(what)],
+        ),
+      );
+      i = j + what.length;
     }
+    if (i < text.length) _texts.add(TextSpan(text: text.substring(i)));
 
     return RichText(
       strutStyle: strutStyle,
@@ -150,16 +169,7 @@ class TextChunkStyling extends StatelessWidget {
       textScaleFactor: textScaleFactor,
       maxLines: maxLines,
       textWidthBasis: textWidthBasis,
-      text: TextSpan(
-        children: combined
-            .map<TextSpan>((e) => TextSpan(
-                  text: e,
-                  style: highlightText.contains(e.toLowerCase())
-                      ? highlightTextStyle
-                      : effectiveTextStyle,
-                ))
-            .toList(),
-      ),
+      text: TextSpan(children: _texts, style: effectiveTextStyle),
     );
   }
 }